requirements.txt
Flask>=2.0.1
Flask-WTF>=0.15.1
Flask-Talisman>=0.8.1
Flask-Limiter>=2.6.0
Pillow>=9.5.0
numpy>=1.24.3
Werkzeug>=2.0.1
WTForms>=2.3.3
opencv-python-headless>=4.5.3.56
setuptools>=65.0.0
gunicorn>=20.1.0
